chore: add fixes
